Move ImmutableString into a separate module
