+++ hyperon-common/src/lib.rs
@@ -10,6 +10,7 @@ pub mod owned_or_borrowed;
 pub mod vecondemand;
 mod flex_ref;
 pub use flex_ref::FlexRef;
+pub mod immutable_string;
 
 use std::collections::HashMap;
