+++ lib/src/metta/runner/str.rs
@@ -1,5 +1,5 @@
 use hyperon_atom::*;
-use hyperon_common::collections::ImmutableString;
+use hyperon_common::immutable_string::ImmutableString;
 use hyperon_atom::serial;
 use hyperon_atom::ConvertingSerializer;
 use unescaper;
